requirements.txt
requests
beautifulsoup4
lxml
pdfplumber
pytesseract
pillow
opencv-python
python-dotenv
SQLAlchemy
psycopg[binary]
pypdf
